Unify VReg and LabelId across codegen backends
Extract VReg and LabelId types from x86_64/mir.rs and aarch64/mir.rs
into a shared vreg.rs module. These types are target-independent since
virtual registers only become physical registers during register
allocation.

Also extract collect_array_scalar_vregs and collect_struct_scalar_vregs
from both cfg_lower.rs files into the shared types.rs module. The
backends now use thin wrapper methods that delegate to the shared
implementations via closures.

Fixes rue-ug59
